requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pymysql
sqlalchemy